Allow `Terminator` to be used in more places
The implementation of `Terminator` is pretty restrictive. There's two
notable places where it falls over:

- It cannot be used if any of the functions you're calling return any
  trait object other than `Box<dyn Error>`. An extremely common type is
  `Box<dyn Error + Send + Sync>`.
- Changing `Result<T, E>` to `Result<T, Terminator>` potentially loses
  context, since you're coercing directly from the expression `?` was
  used on to `Box<dyn Error>`, not `E` first.

The first is a major issue, and cannot be fixed without something
similar to this commit. The second issue could be worked around by
wrapping everything in another function with the error type you want to
convert to, but now we're basically stuck back in 1.25 days. It's also
especially annoying to do that in tests.

This commit addresses these issues by changing `Terminator` to be
generic over any type (with the default being `Box<dyn Error>` to keep
similar behavior as before, and providing an extremely generic `From`
impl that says "we can be converted from any type which can convert to
our error type"

There is a wrinkle here, since the naive impl would overlap with
`impl<T> From<T> for T` in the standard library. To work around this,
I've required that the input type implement `Display`. This is a little
funky, since it means that the type we're converting *from* must
implement `Display`, when we actually only care that the type we're
converting *to* implements it. But it was the best bound I could think
of which does not apply to `Terminator<E>`, and is likely to apply to
every type used as an error.

In the rare case where you're converting from something that doesn't
implement display, you'd need an explicit intermeidate conversion, e.g.
`result.map_err(MyError::from)?`. This seems niche enough that I don't
even think it's worth documenting in the library.